Add appengine logging, remove unsupported metrics
The Go system log package does not print anything to AEStandard log
messages. This change adds support for native AEStandard logging
including line numbers.

As well, this change removes prometheus metrics which cannot be
supported in AEStandard due to internal usage of the 'syscall' package,
which is forbidden in AEStandard.